changed id type in repositories from Integer to Long and added the UserService and UserServiceImplementation classes..also added model mapper Bean

+++ src/main/java/com/example/package_delivery_system/data/repositories/OfficeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfficeRepository extends JpaRepository<Office, Integer> {
+public interface OfficeRepository extends JpaRepository<Office, Long> {
 }
